requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.1.3
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
coverage==7.10.1
cryptography==42.0.5
Deprecated==1.2.18
deprecation==2.1.0
ecdsa==0.19.1
fastapi==0.110.2
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
limits==5.4.0
packaging==25.0
pluggy==1.6.0
postgrest==0.19.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
realtime==2.5.3
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.37.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.4
typing_extensions==4.14.1
uvicorn==0.29.0
websockets==15.0.1
wrapt==1.17.2
